// front-plataforma/src/app/api/generated/model/studentRegisterDTO.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * /auth
 * API for managing user accounts in the platform
 * OpenAPI spec version: 1.0.0
 */

export interface StudentRegisterDTO {
  /** @minLength 1 */
  userName: string;
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  password: string;
}